# Speed-up unit tests