--- a/lib/spack/spack/test/build_systems.py
+++ b/lib/spack/spack/test/build_systems.py
@@ -27,9 +27,9 @@ pytestmark = pytest.mark.skipif(sys.platform == "win32", reason="does not run on
 
 
 @pytest.fixture()
-def concretize_and_setup(mock_concretize):
+def concretize_and_setup(default_mock_concretization):
     def _func(spec_str):
-        s = mock_concretize(spec_str)
+        s = default_mock_concretization(spec_str)
         setup_package(s.package, False)
         return s
 
@@ -95,8 +95,8 @@ class TestTargets(object):
 
 @pytest.mark.usefixtures("config", "mock_packages")
 class TestAutotoolsPackage(object):
-    def test_with_or_without(self, mock_concretize):
-        s = mock_concretize("a")
+    def test_with_or_without(self, default_mock_concretization):
+        s = default_mock_concretization("a")
         options = s.package.with_or_without("foo")
 
         # Ensure that values that are not representing a feature
@@ -127,8 +127,8 @@ class TestAutotoolsPackage(object):
         options = s.package.with_or_without("lorem-ipsum", variant="lorem_ipsum")
         assert "--without-lorem-ipsum" in options
 
-    def test_none_is_allowed(self, mock_concretize):
-        s = mock_concretize("a foo=none")
+    def test_none_is_allowed(self, default_mock_concretization):
+        s = default_mock_concretization("a foo=none")
         options = s.package.with_or_without("foo")
 
         # Ensure that values that are not representing a feature
@@ -138,9 +138,11 @@ class TestAutotoolsPackage(object):
         assert "--without-baz" in options
         assert "--no-fee" in options
 
-    def test_libtool_archive_files_are_deleted_by_default(self, mock_concretize, mutable_database):
+    def test_libtool_archive_files_are_deleted_by_default(
+        self, default_mock_concretization, mutable_database
+    ):
         # Install a package that creates a mock libtool archive
-        s = mock_concretize("libtool-deletion")
+        s = default_mock_concretization("libtool-deletion")
         s.package.do_install(explicit=True)
 
         # Assert the libtool archive is not there and we have
@@ -151,23 +153,25 @@ class TestAutotoolsPackage(object):
         assert libtool_deletion_log
 
     def test_libtool_archive_files_might_be_installed_on_demand(
-        self, mutable_database, monkeypatch, mock_concretize
+        self, mutable_database, monkeypatch, default_mock_concretization
     ):
         # Install a package that creates a mock libtool archive,
         # patch its package to preserve the installation
-        s = mock_concretize("libtool-deletion")
+        s = default_mock_concretization("libtool-deletion")
         monkeypatch.setattr(type(s.package.builder), "install_libtool_archives", True)
         s.package.do_install(explicit=True)
 
         # Assert libtool archives are installed
         assert os.path.exists(s.package.builder.libtool_archive_file)
 
-    def test_autotools_gnuconfig_replacement(self, mock_concretize, mutable_database):
+    def test_autotools_gnuconfig_replacement(self, default_mock_concretization, mutable_database):
         """
         Tests whether only broken config.sub and config.guess are replaced with
         files from working alternatives from the gnuconfig package.
         """
-        s = mock_concretize("autotools-config-replacement +patch_config_files +gnuconfig")
+        s = default_mock_concretization(
+            "autotools-config-replacement +patch_config_files +gnuconfig"
+        )
         s.package.do_install()
 
         with open(os.path.join(s.prefix.broken, "config.sub")) as f:
@@ -182,11 +186,15 @@ class TestAutotoolsPackage(object):
         with open(os.path.join(s.prefix.working, "config.guess")) as f:
             assert "gnuconfig version of config.guess" not in f.read()
 
-    def test_autotools_gnuconfig_replacement_disabled(self, mock_concretize, mutable_database):
+    def test_autotools_gnuconfig_replacement_disabled(
+        self, default_mock_concretization, mutable_database
+    ):
         """
         Tests whether disabling patch_config_files
         """
-        s = mock_concretize("autotools-config-replacement ~patch_config_files +gnuconfig")
+        s = default_mock_concretization(
+            "autotools-config-replacement ~patch_config_files +gnuconfig"
+        )
         s.package.do_install()
 
         with open(os.path.join(s.prefix.broken, "config.sub")) as f:
@@ -250,29 +258,29 @@ spack:
 
 @pytest.mark.usefixtures("config", "mock_packages")
 class TestCMakePackage(object):
-    def test_cmake_std_args(self, mock_concretize):
+    def test_cmake_std_args(self, default_mock_concretization):
         # Call the function on a CMakePackage instance
-        s = mock_concretize("cmake-client")
+        s = default_mock_concretization("cmake-client")
         expected = spack.build_systems.cmake.CMakeBuilder.std_args(s.package)
         assert s.package.builder.std_cmake_args == expected
 
         # Call it on another kind of package
-        s = mock_concretize("mpich")
+        s = default_mock_concretization("mpich")
         assert spack.build_systems.cmake.CMakeBuilder.std_args(s.package)
 
-    def test_cmake_bad_generator(self, monkeypatch, mock_concretize):
-        s = mock_concretize("cmake-client")
+    def test_cmake_bad_generator(self, monkeypatch, default_mock_concretization):
+        s = default_mock_concretization("cmake-client")
         monkeypatch.setattr(type(s.package), "generator", "Yellow Sticky Notes", raising=False)
         with pytest.raises(spack.package_base.InstallError):
             s.package.builder.std_cmake_args
 
-    def test_cmake_secondary_generator(self, mock_concretize):
-        s = mock_concretize("cmake-client")
+    def test_cmake_secondary_generator(self, default_mock_concretization):
+        s = default_mock_concretization("cmake-client")
         s.package.generator = "CodeBlocks - Unix Makefiles"
         assert s.package.builder.std_cmake_args
 
-    def test_define(self, mock_concretize):
-        s = mock_concretize("cmake-client")
+    def test_define(self, default_mock_concretization):
+        s = default_mock_concretization("cmake-client")
 
         define = s.package.define
         for cls in (list, tuple):
@@ -322,8 +330,8 @@ class TestDownloadMixins(object):
             ),
         ],
     )
-    def test_attributes_defined(self, mock_concretize, spec_str, expected_url):
-        s = mock_concretize(spec_str)
+    def test_attributes_defined(self, default_mock_concretization, spec_str, expected_url):
+        s = default_mock_concretization(spec_str)
         assert s.package.urls[0] == expected_url
 
     @pytest.mark.parametrize(
@@ -342,33 +350,33 @@ class TestDownloadMixins(object):
             ("mirror-xorg-broken", r"{0} must define a `xorg_mirror_path` attribute"),
         ],
     )
-    def test_attributes_missing(self, mock_concretize, spec_str, error_fmt):
-        s = mock_concretize(spec_str)
+    def test_attributes_missing(self, default_mock_concretization, spec_str, error_fmt):
+        s = default_mock_concretization(spec_str)
         error_msg = error_fmt.format(type(s.package).__name__)
         with pytest.raises(AttributeError, match=error_msg):
             s.package.urls
 
 
-def test_cmake_define_from_variant_conditional(mock_concretize):
+def test_cmake_define_from_variant_conditional(default_mock_concretization):
     """Test that define_from_variant returns empty string when a condition on a variant
     is not met. When this is the case, the variant is not set in the spec."""
-    s = mock_concretize("cmake-conditional-variants-test")
+    s = default_mock_concretization("cmake-conditional-variants-test")
     assert "example" not in s.variants
     assert s.package.define_from_variant("EXAMPLE", "example") == ""
 
 
-def test_autotools_args_from_conditional_variant(mock_concretize):
+def test_autotools_args_from_conditional_variant(default_mock_concretization):
     """Test that _activate_or_not returns an empty string when a condition on a variant
     is not met. When this is the case, the variant is not set in the spec."""
-    s = mock_concretize("autotools-conditional-variants-test")
+    s = default_mock_concretization("autotools-conditional-variants-test")
     assert "example" not in s.variants
     assert len(s.package.builder._activate_or_not("example", "enable", "disable")) == 0
 
 
-def test_autoreconf_search_path_args_multiple(mock_concretize, tmpdir):
+def test_autoreconf_search_path_args_multiple(default_mock_concretization, tmpdir):
     """autoreconf should receive the right -I flags with search paths for m4 files
     for build deps."""
-    spec = mock_concretize("dttop")
+    spec = default_mock_concretization("dttop")
     aclocal_fst = str(tmpdir.mkdir("fst").mkdir("share").mkdir("aclocal"))
     aclocal_snd = str(tmpdir.mkdir("snd").mkdir("share").mkdir("aclocal"))
     build_dep_one, build_dep_two = spec.dependencies(deptype="build")
@@ -382,11 +390,11 @@ def test_autoreconf_search_path_args_multiple(mock_concretize, tmpdir):
     ]
 
 
-def test_autoreconf_search_path_args_skip_automake(mock_concretize, tmpdir):
+def test_autoreconf_search_path_args_skip_automake(default_mock_concretization, tmpdir):
     """automake's aclocal dir should not be added as -I flag as it is a default
     3rd party dir search path, and if it's a system version it usually includes
     m4 files shadowing spack deps."""
-    spec = mock_concretize("dttop")
+    spec = default_mock_concretization("dttop")
     tmpdir.mkdir("fst").mkdir("share").mkdir("aclocal")
     aclocal_snd = str(tmpdir.mkdir("snd").mkdir("share").mkdir("aclocal"))
     build_dep_one, build_dep_two = spec.dependencies(deptype="build")
@@ -396,9 +404,9 @@ def test_autoreconf_search_path_args_skip_automake(mock_concretize, tmpdir):
     assert spack.build_systems.autotools._autoreconf_search_path_args(spec) == ["-I", aclocal_snd]
 
 
-def test_autoreconf_search_path_args_external_order(mock_concretize, tmpdir):
+def test_autoreconf_search_path_args_external_order(default_mock_concretization, tmpdir):
     """When a build dep is external, its -I flag should occur last"""
-    spec = mock_concretize("dttop")
+    spec = default_mock_concretization("dttop")
     aclocal_fst = str(tmpdir.mkdir("fst").mkdir("share").mkdir("aclocal"))
     aclocal_snd = str(tmpdir.mkdir("snd").mkdir("share").mkdir("aclocal"))
     build_dep_one, build_dep_two = spec.dependencies(deptype="build")
@@ -412,18 +420,18 @@ def test_autoreconf_search_path_args_external_order(mock_concretize, tmpdir):
     ]
 
 
-def test_autoreconf_search_path_skip_nonexisting(mock_concretize, tmpdir):
+def test_autoreconf_search_path_skip_nonexisting(default_mock_concretization, tmpdir):
     """Skip -I flags for non-existing directories"""
-    spec = mock_concretize("dttop")
+    spec = default_mock_concretization("dttop")
     build_dep_one, build_dep_two = spec.dependencies(deptype="build")
     build_dep_one.prefix = str(tmpdir.join("fst"))
     build_dep_two.prefix = str(tmpdir.join("snd"))
     assert spack.build_systems.autotools._autoreconf_search_path_args(spec) == []
 
 
-def test_autoreconf_search_path_dont_repeat(mock_concretize, tmpdir):
+def test_autoreconf_search_path_dont_repeat(default_mock_concretization, tmpdir):
     """Do not add the same -I flag twice to keep things readable for humans"""
-    spec = mock_concretize("dttop")
+    spec = default_mock_concretization("dttop")
     aclocal = str(tmpdir.mkdir("prefix").mkdir("share").mkdir("aclocal"))
     build_dep_one, build_dep_two = spec.dependencies(deptype="build")
     build_dep_one.external_path = str(tmpdir.join("prefix"))