… lib/spack/spack/test/ci.py
… lib/spack/spack/test/database.py
… lib/spack/spack/test/git_fetch.py
… lib/spack/spack/test/packaging.py
… lib/spack/spack/test/spec_semantics.py
… lib/spack/spack/test/url_fetch.py