Results on my machine:

* **Spack:** 0.19.0.dev0 (fd5f8a666af292e2513ebe206bdb60f1387bf6bd)
* **Python:** 3.8.10
* **Platform:** linux-ubuntu20.04-icelake
* **Concretizer:** clingo


| Test file  | Time on develop | Time in this PR | Speedup |
| ----------- | ---------------------- | -------------------- | ------------ |
|  lib/spack/spack/test/architecture.py  | 18.18s  | 3.32s | **5.48x** |
| lib/spack/spack/test/spec_semantics.py  | 87.96s  | 11.18s | **7.87x** |
| lib/spack/spack/test/ci.py  | 24.79s  | 3.83s | **6.49x** |
| lib/spack/spack/test/url_fetch.py  | 53.32s  | 12.99s | **4.10x** |
| lib/spack/spack/test/git_fetch.py  | 26.56s  | 11.89s | **2.33x** |
| lib/spack/spack/test/directory_layout.py  | 17.66s  | 3.67s | **4.81x** |
| lib/spack/spack/test/build_systems.py  | 18.85s  | 12.86s | **1.47x** |
| lib/spack/spack/test/packaging.py  | 4.66s  | 2.65s | 1.76x |
| lib/spack/spack/test/database.py  | 9.34s  | 8.85s | 1.06x |



Modifications:
- [x] Avoid concretization in a few tests in `architecture.py`
- [x] Use mock packages everywhere in `spec_semantics`, reduce re-try in a test
- [x] Cache recurring concretizations in other places