@trws 2022-11-10
Given the notes and how it looks like this works, should `mock_concretize` perhaps be named something like `concretize_and_cache`?  Or maybe `cache_concrete_mock` if you meant to be invoking that it's concretizing a mock package specifically?