@alalazo 2022-11-10
@trws I have no real preferences on the name. I called it `mock_concretize` since the caveat with that fixture is that it implies using the `config` and `mock_packages` fixtures. This is an important bit, as the fixture might return wrong results if used in tests that need to modify the configuration or the `builtin.mock` repository.[^1] So, if there is agreement on a better name I'm fine with that.

[^1]: The full story is that the fixture can be used in those tests only _before_ applying the modifications.