@becker33 2022-11-10
Doesn't the `config` fixture already imply that we won't modify the config, as in that case we have to use the `mutable_config` fixture? Similarly, I think we have another fixture that @scheibelp wrote for tests that need to modify a repo, and we don't use `builtin.mock` for that.

I do think this would be a little clearer for readers with a name like `caching_test_concretizer`, but (A) I'm more sold on the idea of a clearer name, but I don't have a good idea for one, and (B) I don't think it is make-or-break for the PR.