@alalazo 2022-11-10
> Doesn't the config fixture already imply that we won't modify the config, as in that case we have to use the mutable_config fixture? 

It implies we don't modify the existing scopes. But in some tests we add more scopes and then concretize. `mock_concretize` shouldn't be used in those cases.