@alalazo 2022-11-10
What do people think about `fast_concretize`? I wouldn't use "cache" since it ties an implementation detail to the name of the function (and we might decide to change that detail at some point in the future).