@trws 2022-11-11
It isn't a fast concretize though, it's a cached concretize.  I don't think that's an implementation detail, it's the intended behavior, and it matters for some things like testing something after the first concretization or second which we have tests for.  Either way, honestly I want this merged a whole lot more than I want to worry about the name.  The point about it being unusable after modifications are made to config or mock_packages does seem important, but I'm not sure how to incorporate that.  If you don't like cached, how about `default_concretization` or maybe append `_of_mock`? I kinda like that because it implies you get the default concretization regardless of what you might set without saying explicitly that it uses a cache, the extension would make it clear it uses a mock package repo by default too.