@haampie 2022-11-11
For me this reduces runtime by 13%, so a solid improvement